split resolver behavior ref helpers
