+++ src/resolver/metadata_helpers.rs
@@ -1,11 +1,13 @@
-use crate::ast::{
-    behavior_type_args_match_target_params, AstType, BehaviorMethod, EnumVariant, Param,
-    StructField, TypeParam,
-};
+use crate::ast::{AstType, BehaviorMethod, EnumVariant, Param, StructField, TypeParam};
 
 use super::symbol_table::{
     BehaviorMethodTypeMetadata, MethodSignatureMetadata, TypeParameterBoundMetadata,
     TypeParameterBoundRefMetadata, ValueSignatureMetadata,
+};
+
+mod behavior_refs;
+pub(super) use behavior_refs::{
+    behavior_ref_display, resolver_behavior_impl_method_key_with_target_args, resolver_method_key,
 };
 
 fn resolver_return_type_name(return_type: &Option<AstType>) -> String {
@@ -97,93 +99,6 @@ fn type_param_bound_display(type_param: &TypeParam) -> Option<String> {
     })
 }
 
-pub(super) fn behavior_ref_display(behavior: &str, type_args: &[AstType]) -> String {
-    if type_args.is_empty() {
-        behavior.to_string()
-    } else {
-        format!(
-            "{}<{}>",
-            behavior,
-            type_args
-                .iter()
-                .map(AstType::display_name)
-                .collect::<Vec<_>>()
-                .join(", ")
-        )
-    }
-}
-
-pub(super) fn resolver_method_key(type_name: &str, method_name: &str) -> String {
-    format!("{type_name}.{method_name}")
-}
-
-pub(super) fn resolver_behavior_impl_method_key(
-    type_name: &str,
-    method_name: &str,
-    behavior: &str,
-    behavior_type_args: &[AstType],
-) -> String {
-    if behavior_type_args.is_empty() {
-        return resolver_method_key(type_name, method_name);
-    }
-
-    format!(
-        "{}__{}",
-        resolver_method_key(type_name, method_name),
-        behavior_ref_symbol_suffix(behavior, behavior_type_args)
-    )
-}
-
-pub(super) fn resolver_behavior_impl_method_key_with_target_args(
-    type_name: &str,
-    method_name: &str,
-    behavior: &str,
-    behavior_type_args: &[AstType],
-    target_type_args: &[AstType],
-) -> String {
-    if target_type_args.is_empty() {
-        return resolver_behavior_impl_method_key(
-            type_name,
-            method_name,
-            behavior,
-            behavior_type_args,
-        );
-    }
-
-    if behavior_type_args_match_target_params(behavior_type_args, target_type_args) {
-        format!(
-            "{}__{}",
-            resolver_method_key(type_name, method_name),
-            behavior
-        )
-    } else {
-        resolver_behavior_impl_method_key(type_name, method_name, behavior, behavior_type_args)
-    }
-}
-
-pub(super) fn behavior_ref_symbol_suffix(behavior: &str, type_args: &[AstType]) -> String {
-    let mut parts = vec![sanitize_symbol_part(behavior)];
-    parts.extend(
-        type_args
-            .iter()
-            .map(AstType::display_name)
-            .map(|name| sanitize_symbol_part(&name)),
-    );
-    parts.join("_")
-}
-
-fn sanitize_symbol_part(name: &str) -> String {
-    let mut out = String::new();
-    for ch in name.chars() {
-        if ch.is_ascii_alphanumeric() {
-            out.push(ch);
-        } else {
-            out.push('_');
-        }
-    }
-    out
-}
-
 pub(super) fn resolver_field_types(fields: &[StructField]) -> Vec<(String, AstType, String)> {
     fields
         .iter()
@@ -232,6 +147,9 @@ pub(super) fn resolver_behavior_method_types(
         })
         .collect()
 }
+
+#[cfg(test)]
+use behavior_refs::resolver_behavior_impl_method_key;
 
 #[cfg(test)]
 mod tests;
